requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
line-bot-sdk>=3.0.0
psycopg2-binary>=2.9.0
pg8000>=1.31.0
psutil>=5.9.0
